Adds readme, webpack, etc. Button layout wip;
Used laravel mix as a starter for a webpack layout because I love
working with browserSync and automatic reloads, hate to work without.

I started a base readme and will fill out later.

Power went out, on 2% battery and it's nearly 11pm. I'm somewhat close
but also at a loss on how to get the stars and subtle gloss but
determined to get it. But pushing for now while I have 2%. Until
tommorrow. 1% now. Typing fast.